Catch new error

+++ src/util/research.py
@@ -491,6 +491,9 @@ class Artifact:
                         )
 
                     # Copy it. Works for files or directories.
+                    if not os.path.exists(obj): 
+                        raise FileNotFoundError(obj) 
+
                     if obj.is_file():
                         shutil.copy(obj, path)
                     elif obj.is_dir():
